backend/requirements.txt
# SmartAP Phase 1 - Core Intake Engine Dependencies
# FastAPI Backend
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12

# AI/Agent Framework (--pre flag required during preview)
agent-framework-azure-ai
openai>=1.50.0

# Data Validation & Models
pydantic>=2.9.0
pydantic-settings>=2.5.0
email-validator>=2.0.0

# PDF Processing (placeholder for Foxit SDK integration)
# foxit-pdf-sdk  # Uncomment when Foxit SDK is available
pypdf>=5.0.0  # Fallback for basic PDF operations

# Utilities
python-dotenv>=1.0.1
httpx>=0.27.0
aiofiles>=24.1.0

# Development & Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0

# Database (Phase 2)
sqlalchemy[asyncio]>=2.0.25
alembic>=1.13.0
aiosqlite>=0.19.0  # Development database
asyncpg>=0.29.0  # Production PostgreSQL
psycopg2-binary>=2.9.0  # PostgreSQL driver

# Redis Cache (Phase 2.6)
redis[hiredis]>=5.0.0

# Fuzzy Matching & Similarity (Phase 2)
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.25.0

# Agent Orchestration (Phase 2.4)
langgraph>=0.2.0
langchain-core>=0.3.0

# ERP Integration (Phase 4.3)
apscheduler>=3.10.0

# Logging (Phase 2.6)
python-json-logger>=2.0.7

# Monitoring (optional)
# prometheus-client>=0.19.0
# sentry-sdk[fastapi]>=1.39.0

